Add files via upload
Created a script named extract_text_from_rotten_tomatoes.py

This script defines a function that accepts the URL to a movie on RottenTomatoes. It then creates a text file that includes the following information for each review in the first 2 review pages for the movie:

- the name of the critic 

- the rating. The rating should be 'rotten', 'fresh', or 'NA' if the review doesn't have a rating.

- the source (e.g 'New York Daily News') of the review. Is 'NA' if the review doesn't have a source.

- the text of the review. Is 'NA' if the review doesn't have text.

- the date of the review. Is 'NA' if the review doesn't have a date.

The file includes one line for each review. The reviews in the file appear in the same order as they do on the website. The 5 values that you write for each movie is written in the order listed above. The 5 values are separated by a TAB.